refactor: move TTL input to HomeFragment
Move TTL input from content_main.xml to fragment_home.xml

This commit removes the LinearLayout with EditText and two Buttons from content_main.xml and adds it to fragment_home.xml. The change makes more sense since the TTL input is related to the HomeFragment and not the entire app.