docs: Add alternatives to usage section
Add alternatives to the Java Android TTL Manager app in the usage section. Specifically, add a reference to TTL Editor, a simple graphical front-end for iptables to change the TTL of packets sent over a given network interface, available on F-Droid, and encourage users to write or make PRs for other open-source alternatives.